+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Add inPlace option to write hashes in place. Add updateExistingHashes to control if existing hashes are updated.
+
 ## [1.0.2] - 2024-11-22
 
 ### Added
@@ -13,4 +19,5 @@
 
 - Increase version
 
+[Unreleased]: https://github.com/inlavigo/gg_json_hash/compare/1.0.2...HEAD
 [1.0.2]: https://github.com/inlavigo/gg_json_hash/tag/%tag